refactor: rename contexts to stores
